setting `body` direction to `rtl` in theme
In a multi-lingual projects it is not easy to set `dir` attribute of `body` element when the user change the language which cause the direction change ( `rtl` <-> `ltr` ), becase in react practices, the `body` element is normally out of the react scope. I checked that it works if we use `@global` styling and change `direction` style of the `body` element. This way the developer can set the `direction` depending on the selected language on the theme creation code.

One thing that makes me suspicious is this sentence "Make sure the `dir` attribute is set on the body, otherwise native components will break", does it work if set `direction` style and not `dir` attribute?